Integer tensor fixes

+++ cr-examples/onnx/src/main/java/oracle/code/onnx/Tensor.java
@@ -107,11 +107,11 @@ public class Tensor<T> extends OnnxNumber {
         return ofShape(arena, new long[]{values.length}, values);
     }
 
-    public static Tensor<Long> ofFlat(int... values) {
+    public static Tensor<Integer> ofFlat(int... values) {
         return ofShape(new long[]{values.length}, values);
     }
 
-    public static Tensor<Long> ofFlat(Arena arena, int... values) {
+    public static Tensor<Integer> ofFlat(Arena arena, int... values) {
         return ofShape(arena, new long[]{values.length}, values);
     }
 
@@ -147,11 +147,11 @@ public class Tensor<T> extends OnnxNumber {
         return new Tensor(arena, arena.allocateFrom(ValueLayout.JAVA_BYTE, values), ElementType.UINT8, shape);
     }
 
-    public static Tensor<Long> ofShape(long[] shape, int... values) {
+    public static Tensor<Integer> ofShape(long[] shape, int... values) {
         return ofShape(Arena.ofAuto(), shape, values);
     }
 
-    public static Tensor<Long> ofShape(Arena arena, long[] shape, int... values) {
+    public static Tensor<Integer> ofShape(Arena arena, long[] shape, int... values) {
         return new Tensor(arena, arena.allocateFrom(ValueLayout.JAVA_INT, values), ElementType.INT32, shape);
     }
